# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Twitch Bot API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface BotStatus {
  running: boolean;
  /** @nullable */
  channel: string | null;
  uptime_seconds: number;
  messages_sent: number;
  /** @nullable */
  last_action: string | null;
}

export interface BotSettings {
  id: number;
  channel_name: string;
  bot_username: string;
  twitch_oauth_token?: string;
  openai_api_key?: string;
  gemini_api_key?: string;
  personality: string;
  min_delay_seconds: number;
  max_delay_seconds: number;
  cooldown_seconds: number;
  respond_to_chat: boolean;
  vision_enabled: boolean;
  speech_enabled: boolean;
  /** @nullable */
  active_hours_start: number | null;
  /** @nullable */
  active_hours_end: number | null;
  created_at: string;
  updated_at: string;
}

export interface BotSettingsUpdate {
  channel_name?: string;
  bot_username?: string;
  twitch_oauth_token?: string;
  openai_api_key?: string;
  gemini_api_key?: string;
  personality?: string;
  min_delay_seconds?: number;
  max_delay_seconds?: number;
  cooldown_seconds?: number;
  respond_to_chat?: boolean;
  vision_enabled?: boolean;
  speech_enabled?: boolean;
  /** @nullable */
  active_hours_start?: number | null;
  /** @nullable */
  active_hours_end?: number | null;
}

export type LogEntryType = typeof LogEntryType[keyof typeof LogEntryType];


export const LogEntryType = {
  vision: 'vision',
  speech: 'speech',
  chat: 'chat',
  decision: 'decision',
  message_sent: 'message_sent',
  error: 'error',
} as const;

export interface LogEntry {
  id: number;
  type: LogEntryType;
  content: string;
  /** @nullable */
  metadata?: string | null;
  created_at: string;
}

export type BotMessageTriggerType = typeof BotMessageTriggerType[keyof typeof BotMessageTriggerType];


export const BotMessageTriggerType = {
  vision: 'vision',
  speech: 'speech',
  chat_reply: 'chat_reply',
  scheduled: 'scheduled',
} as const;

export interface BotMessage {
  id: number;
  channel: string;
  message: string;
  trigger_type: BotMessageTriggerType;
  /** @nullable */
  context_summary?: string | null;
  created_at: string;
}

export type ChatPatternPatternType = typeof ChatPatternPatternType[keyof typeof ChatPatternPatternType];


export const ChatPatternPatternType = {
  reaction: 'reaction',
  hype: 'hype',
  question: 'question',
  joke: 'joke',
  emote_combo: 'emote_combo',
  game_specific: 'game_specific',
  cs2_callout: 'cs2_callout',
  russian_slang: 'russian_slang',
} as const;

export interface ChatPattern {
  id: number;
  source_channel: string;
  pattern_type: ChatPatternPatternType;
  content: string;
  frequency: number;
  language?: string;
  game?: string;
  created_at: string;
}

export interface LearnRequest {
  channel: string;
  message_count?: number;
}

export interface LearnResponse {
  success: boolean;
  patterns_found: number;
  channel: string;
}

export interface BulkLearnRequest {
  /** Custom channel list. If omitted, uses preset Russian CS2 streamers. */
  channels?: string[];
  message_count_per_channel?: number;
}

export interface BulkLearnResponse {
  success: boolean;
  channels_queued: string[];
  total_channels: number;
}

export interface BotStats {
  total_messages_sent: number;
  messages_today: number;
  total_patterns_learned: number;
  channels_learned_from: number;
  uptime_hours: number;
  avg_response_delay_seconds: number;
  /** @nullable */
  most_common_trigger: string | null;
}

export interface DetectLiveRequest {
  channels?: string[];
  window_ms?: number;
}

export interface CollectChatRequest {
  message_count?: number;
}

export type StreamerPresetCategory = typeof StreamerPresetCategory[keyof typeof StreamerPresetCategory];


export const StreamerPresetCategory = {
  entertainment: 'entertainment',
  pro: 'pro',
  variety: 'variety',
} as const;

export interface StreamerPreset {
  channel: string;
  displayName: string;
  description: string;
  category: StreamerPresetCategory;
  priority: number;
}

export interface StreamerFile {
  channel: string;
  display_name?: string;
  category?: string;
  description?: string;
  has_raw: boolean;
  has_patterns: boolean;
  has_analysis: boolean;
  /** @nullable */
  collected_at?: string | null;
  total_messages: number;
  pattern_count: number;
}

export interface ChannelActivity {
  channel: string;
  message_count: number;
  is_live: boolean;
  messages_per_minute: number;
}

export interface LiveDetectionResult {
  results: ChannelActivity[];
  detected_at: string;
}

export interface PatternEntry {
  content: string;
  type: string;
  frequency: number;
  language: string;
  game: string;
}

export interface PatternTypeStat {
  type: string;
  count: number;
  percent: number;
}

export interface StreamerStats {
  total_messages: number;
  unique_phrases: number;
  ru_ratio: number;
  en_ratio: number;
  mixed_ratio: number;
  cs2_ratio: number;
  avg_message_length: number;
  top_pattern_types: PatternTypeStat[];
}

export interface StreamerAnalysis {
  channel: string;
  analyzed_at: string;
  stats: StreamerStats;
  top_phrases: PatternEntry[];
  sample_messages: string[];
  style_notes: string[];
}

export type GetLogsParams = {
limit?: number;
type?: GetLogsType;
};

export type GetLogsType = typeof GetLogsType[keyof typeof GetLogsType];


export const GetLogsType = {
  vision: 'vision',
  speech: 'speech',
  chat: 'chat',
  decision: 'decision',
  message_sent: 'message_sent',
  error: 'error',
} as const;

export type GetMessagesParams = {
limit?: number;
};

export type GetPatternsParams = {
limit?: number;
language?: string;
game?: string;
};

export type ClearPatterns200 = {
  deleted: number;
};

export type CollectStreamerChat200 = {
  started: boolean;
  channel: string;
  message_count: number;
};