# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Twitch Bot API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /streamers/{channel}/collect:
    post:
      operationId: collectStreamerChat
      tags: [streamers]
      summary: Start collecting chat from a specific streamer channel
      parameters:
        - name: channel
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CollectChatRequest"
      responses:
        "200":
          description: Collection started
          content:
            application/json:
              schema:
                type: object
                required: [started, channel, message_count]
                properties:
                  started:
                    type: boolean
                  channel:
                    type: string
                  message_count:
                    type: integer
  /streamers/{channel}/analysis:
    get:
      operationId: getStreamerAnalysis
      tags: [streamers]
      summary: Get analysis for a specific streamer
      parameters:
        - name: channel
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Streamer analysis
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StreamerAnalysis"
        "404":
          description: No data found
  /streamers/{channel}/patterns:
    get:
      operationId: getStreamerPatterns
      tags: [streamers]
      summary: Get saved patterns for a streamer from file
      parameters:
        - name: channel
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Streamer patterns
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PatternEntry"
        "404":
          description: No data found
components:
  schemas:
...
    CollectChatRequest:
      type: object
      properties:
        message_count:
          type: integer
          default: 500
...
    PatternEntry:
      type: object
      required: [content, type, frequency, language, game]
      properties:
        content:
          type: string
        type:
          type: string
        frequency:
          type: integer
        language:
          type: string
        game:
          type: string
    PatternTypeStat:
      type: object
      required: [type, count, percent]
      properties:
        type:
          type: string
        count:
          type: integer
        percent:
          type: integer
    StreamerStats:
      type: object
      required: [total_messages, unique_phrases, ru_ratio, en_ratio, mixed_ratio, cs2_ratio, avg_message_length, top_pattern_types]
      properties:
        total_messages:
          type: integer
        unique_phrases:
          type: integer
        ru_ratio:
          type: integer
        en_ratio:
          type: integer
        mixed_ratio:
          type: integer
        cs2_ratio:
          type: integer
        avg_message_length:
          type: integer
        top_pattern_types:
          type: array
          items:
            $ref: "#/components/schemas/PatternTypeStat"
    StreamerAnalysis:
      type: object
      required: [channel, analyzed_at, stats, top_phrases, sample_messages, style_notes]
      properties:
        channel:
          type: string
        analyzed_at:
          type: string
        stats:
          $ref: "#/components/schemas/StreamerStats"
        top_phrases:
          type: array
          items:
            $ref: "#/components/schemas/PatternEntry"
        sample_messages:
          type: array
          items:
            type: string
        style_notes:
          type: array
          items:
            type: string