# lib/api-zod/src/generated/types/patternEntry.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Twitch Bot API
 * OpenAPI spec version: 0.1.0
 */

export interface PatternEntry {
  content: string;
  type: string;
  frequency: number;
  language: string;
  game: string;
}