# lib/api-zod/src/generated/types/patternTypeStat.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Twitch Bot API
 * OpenAPI spec version: 0.1.0
 */

export interface PatternTypeStat {
  type: string;
  count: number;
  percent: number;
}