# lib/api-zod/src/generated/types/streamerAnalysis.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Twitch Bot API
 * OpenAPI spec version: 0.1.0
 */
import type { PatternEntry } from './patternEntry';
import type { StreamerStats } from './streamerStats';

export interface StreamerAnalysis {
  channel: string;
  analyzed_at: string;
  stats: StreamerStats;
  top_phrases: PatternEntry[];
  sample_messages: string[];
  style_notes: string[];
}