# lib/api-zod/src/generated/types/streamerStats.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Twitch Bot API
 * OpenAPI spec version: 0.1.0
 */
import type { PatternTypeStat } from './patternTypeStat';

export interface StreamerStats {
  total_messages: number;
  unique_phrases: number;
  ru_ratio: number;
  en_ratio: number;
  mixed_ratio: number;
  cs2_ratio: number;
  avg_message_length: number;
  top_pattern_types: PatternTypeStat[];
}